moved DummyForecaster to utils
